refactoring to satisfy same location filter for the other datasources
